move annotate action to annotations controller
