+++ app/controllers/projects_controller.rb
@@ -2,7 +2,6 @@ class ProjectsController < ApplicationController
   include ServiceRoles
 
   before_action :set_project, only: [
-    :annotate,
     :bootstrap,
     :check_interfaces,
     :check_problem_identifiers,
@@ -102,16 +101,6 @@ class ProjectsController < ApplicationController
     else
       redirect_bootstrap_with_flash
     end
-  end
-
-  # GET /projects/1/annotate
-  def annotate
-    if @project.annotation_documents.empty?
-      flash[:alert] = I18n.t 'projects.annotate.errors.no-annotation-documents'
-      redirect_to project_path(@project)
-    end
-
-    @container_class = Rails.configuration.x.dalphi['annotation-interface']['container-class-name']
   end
 
   # GET /projects/1/bootstrap
