+++ app/helpers/annotations_helper.rb
@@ -1,0 +1,2 @@
+module AnnotationsHelper
+end
