+++ app/views/shared/_sidebar.slim
@@ -15,7 +15,7 @@
     - if user_signed_in?
       ul.nav
         - active_class = ''
-        - active_class = 'active' if %w(projects raw_data).include?(params[:controller])
+        - active_class = 'active' if %w(projects raw_data annotations).include?(params[:controller])
         li class="#{active_class}"
           a href="#{projects_path}"
             = t('activerecord.models.project.other')
