+++ spec/controllers/annotations_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe AnnotationsController, type: :controller do
+
+end
